# docs/swagger.yaml
definitions:
  go-enterprise-api_internal_application_auth.LoginResult:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
      user:
        additionalProperties: {}
        type: object
    type: object
  go-enterprise-api_internal_domain_audit.Entry:
    properties:
      action:
        type: string
      created_at:
        type: string
      ip_address:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      method:
        type: string
      new_data:
        additionalProperties: {}
        type: object
      old_data:
        additionalProperties: {}
        type: object
      path:
        type: string
      request_id:
        type: string
      resource:
        type: string
      resource_id:
        type: string
      user_agent:
        type: string
      user_id:
        type: integer
    type: object
...
  go-enterprise-api_internal_interface_httpapi.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  go-enterprise-api_internal_interface_httpapi.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorBody'
    type: object
...
  internal_interface_http_handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
host: localhost:18080
info:
  contact:
    name: GitOps
  description: 'Enterprise boilerplate: Dynamic Rate Limit, Soft Delete, Audit Logs,
    RBAC, JWT + Redis sessions, Microsoft Entra (OIDC). All requests pass through
    the Nginx gateway (X-Gateway-Token handshake) — public endpoints are also rate-limited
    per IP at the gateway.'
  title: Go Enterprise API
  version: "1.0"
paths:
  /api/v1/audit-logs:
    get:
      description: 'Paginated, filterable trail of WHO/WHAT/DATA/WHEN/FROM. Filters:
        user_id, action, resource, resource_id, request_id.'
      parameters:
      - description: Filter by user
        in: query
        name: user_id
        type: integer
      - description: Filter by action (LOGIN, USER_CREATED, …)
        in: query
        name: action
        type: string
      - description: Filter by resource (users, roles, …)
        in: query
        name: resource
        type: string
      - description: Filter by resource id
        in: query
        name: resource_id
        type: string
      - description: Filter by request id
        in: query
        name: request_id
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go-enterprise-api_internal_domain_audit.Entry'
            type: array
      security:
      - BearerAuth: []
      summary: Search audit logs
      tags:
      - audit
    post:
      consumes:
      - application/json
      description: Records a raw audit entry. The authenticated principal, request
        id, method and path are attached automatically.
      parameters:
      - description: Audit entry
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-enterprise-api_internal_domain_audit.Entry'
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Write a manual audit entry
      tags:
      - audit
  /api/v1/auth/entra/callback:
    get:
      description: Exchanges the authorization code for a verified profile, JIT-provisions
        the user with the default role and issues a session. Returns the standard
        LoginResult JSON (or a fragment redirect when OAUTH_FRONTEND_URL is set).
      parameters:
      - description: Authorization code from Microsoft
        in: query
        name: code
        required: true
        type: string
      - description: CSRF state returned from /entra/login
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_application_auth.LoginResult'
        "400":
          description: missing/invalid code or state
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "502":
          description: exchange with Microsoft failed
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      summary: Entra OIDC callback
      tags:
      - auth
  /api/v1/auth/entra/login:
    get:
      description: Issues a CSRF state and 302-redirects the browser to Microsoft.
        Unconfigured when OAUTH_CLIENT_ID/OAUTH_TENANT_ID are empty → 503 so the SPA
        can fall back to the local form.
      responses:
        "302":
          description: Redirect to login.microsoftonline.com
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      summary: Start Microsoft Entra (OIDC) login
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates with the local provider and returns a JWT access
        token. Route-limited by the dynamic rule POST:/api/v1/auth/login (10 req/min
        by default).
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_application_auth.LoginResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      summary: Login with email/password
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      description: Revokes the Redis session, invalidating the JWT immediately.
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout / revoke session
      tags:
      - auth
  /api/v1/auth/me:
    get:
      description: Returns the authenticated user with their effective roles.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Current user profile
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Provide a JWT access token obtained from /api/v1/auth/login or the
      Entra OIDC flow as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"